# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API documentation for the Ticket Place App",
        "contact": {
            "email": "ticket.place@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://apid.c4m.mg/ticket-place-app/public",
            "description": "Production Server"
        },
        {
            "url": "http://localhost:8000",
            "description": "Local Server"
        }
    ],
    "paths": {
...
        "/api/cart": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "summary": "Get The user's cart",
                "description": "return user's cart",
                "operationId": "659a9669ea36c97dc1330603eeadc37b",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "aucun résultat trouvé"
                    },
                    "403": {
                        "description": "action forbiden"
                    },
                    "401": {
                        "description": "action unauthorized"
                    },
                    "500": {
                        "description": "erreur serveur"
                    }
                }
            }
        },
        "/api/cart/add": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Add items to cart",
                "description": "Add one or many place event to cart.",
                "operationId": "408de867ef91c8a1c1298be6461a5e3c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "type_place_id": {
                                                    "description": "ID place à ajouter",
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "nombre": {
                                                    "description": "Le nombre de places à ajouter",
                                                    "type": "integer",
                                                    "example": "1"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "minItems": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ajoutés au panier avec succès"
                    },
                    "400": {
                        "description": "Données invalides"
                    },
                    "404": {
                        "description": "Événement non trouvé"
                    }
                }
            }
        },
        "/api/cart/update": {
            "put": {
                "tags": [
                    "Cart"
                ],
                "summary": "Update cart",
                "description": "Update one or many place event of cart.",
                "operationId": "e5b1ca74b4593556a933c9443d5fe0f9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "type_place_id": {
                                                    "description": "ID place à modifier",
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "nombre": {
                                                    "description": "Le nombre de places à modifier",
                                                    "type": "integer",
                                                    "example": "2"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "minItems": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Modification panier succès"
                    },
                    "400": {
                        "description": "Données invalides"
                    },
                    "401": {
                        "description": "Action non authoriser"
                    },
                    "403": {
                        "description": "Action interdit"
                    },
                    "404": {
                        "description": "Événement non trouvé"
                    }
                }
            }
        },
        "/api/cart/clear": {
            "delete": {
                "tags": [
                    "Cart"
                ],
                "summary": "Clear The user's cart",
                "description": "Remove all item of user's cart",
                "operationId": "c2cf94402630acad1cc6d317ec806693",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "403": {
                        "description": "action forbiden"
                    },
                    "401": {
                        "description": "action unauthorized"
                    },
                    "500": {
                        "description": "erreur serveur"
                    }
                }
            }
        },
        "/api/cart/remove/item": {
            "delete": {
                "tags": [
                    "Cart"
                ],
                "summary": "Remove Item",
                "description": "Remove Item in the user's cart.",
                "operationId": "99c07bf01ac5d6a150c741000a5ae399",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "type_place_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "operation succès"
                    },
                    "400": {
                        "description": "Données invalides"
                    },
                    "401": {
                        "description": "Action non authoriser"
                    },
                    "403": {
                        "description": "Action interdit"
                    },
                    "404": {
                        "description": "Type Place non trouvé"
                    }
                }
            }
        },
        "/api/cart/pay": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Pay the cart",
                "description": "Process Cart Payment",
                "operationId": "acf61b071163a06cbdfe6409d21985a7",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "aucun résultat trouvé"
                    },
                    "403": {
                        "description": "action forbiden"
                    },
                    "401": {
                        "description": "action unauthorized"
                    },
                    "500": {
                        "description": "erreur serveur"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}